requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
requests==2.31.0
customtkinter==5.2.2
pycryptodome==3.20.0
ecdsa==0.18.0
pillow